requirements.txt
streamlit
python-dotenv
yt-dlp
ffmpeg-python
pydub

openai-whisper
transformers
torch
sentencepiece

tqdm
